// src/app/persona/form/Models/sripersona.model.ts
export class SriPersona
{
	sriId: any;
	sriApellido: String;
	sriCanton: String;
	sriCedula: String;
	sriCorreo: String;
	sriDirRecidencia: String;
	sriDiscapacidad: any;
	sriEstadoCivil: String;
	sriEtnia: String;
	sriFechaNaciemineto: Date;
	sriGrupoSanguineo: String;
	sriNacionalidad: String;
	sriNombre: String;
	sriNommadre: String;
	sriNompadre: String;
	sriOcupacion: String;
	sriParr: String;
	sriParroquia: String;
	sriProvincia: String;
	sriSexo: String;
	sriTelefono: String;
	sriTelpadres: String;
	sriTipoDiscapacidad: String;
}