// src/app/persona/form/Models/person.model.ts
export class FdiPersona
{
    prsId: any;
	prsApellido: String;
	prsCanton: String;
	prsCedula: String;
	prsCorreo: String;
	prsDirRecidencia: String;
	prsDiscapacidad: any;
	prsEstadoCivil: String;
	prsEtnia: String;
	prsFechaNaciemineto: Date;
	prsGrupoSanguineo: String;
	prsNacionalidad: String;
	prsNombre: String;
	prsNommadre: String;
	prsNompadre: String;
	prsOcupacion: String;
	prsParroquia: String;
	prsProvincia: String;
	prsSexo: String;
	prsTelefono: String;
	prsTelpadres: String;
	prsTipoDiscapacidad: String;
}